requirements.txt
opencv-python
pytesseract
panda
numpy
pillow
